requirements.txt
# Core
requests
python-dotenv
python-binance

# FastAPI backend
fastapi
uvicorn
